requirements.txt
discord.py
python-dotenv
aiohttp
aiosqlite
passlib
